+++ ksmartteam01/src/main/java/anypet/ksmartteam01/admin/controller/AdminGoodsListController.java
@@ -1,5 +1,18 @@
 package anypet.ksmartteam01.admin.controller;
 
-public class AdminGoodsListController {
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 
+@Controller
+@RequestMapping("/goods")
+public class AdminGoodsListController {
+	
+	@GetMapping("/goodsCategoryList")
+	public String goodsLargeCategoryList() {
+		
+		return "/user/goods/goodsCategoryList";
+	}
 }
